+++ addons/storyshots/storyshots-core/src/frameworks/angular/loader.ts
@@ -7,17 +7,17 @@ import { StoryshotsOptions } from '../../api/StoryshotsOptions';
 
 function setupAngularJestPreset() {
   // Needed to prevent "Zone.js has detected that ZoneAwarePromise `(window|global).Promise` has been overwritten."
-  require.requireActual('core-js');
-  require.requireActual('core-js/modules/es.promise');
-  // require.requireActual('core-js/es6/reflect');
-  // require.requireActual('core-js/es7/reflect');
+  jest.requireActual('core-js');
+  jest.requireActual('core-js/modules/es.promise');
+  // jest.requireActual('core-js/es6/reflect');
+  // jest.requireActual('core-js/es7/reflect');
 
   // Angular + Jest + Storyshots = Crazy Shit:
   // We need to require 'jest-preset-angular/setupJest' before any storybook code
   // is running inside jest -  one of the things that `jest-preset-angular/setupJest` does is
   // extending the `window.Reflect` with all the needed metadata functions, that are required
   // for emission of the TS decorations like 'design:paramtypes'
-  require.requireActual('jest-preset-angular/setupJest');
+  jest.requireActual('jest-preset-angular/setupJest');
 }
 
 function test(options: StoryshotsOptions): boolean {
@@ -29,13 +29,13 @@ function test(options: StoryshotsOptions): boolean {
 function load(options: StoryshotsOptions) {
   setupAngularJestPreset();
 
-  const storybook = require.requireActual('@storybook/angular');
+  const storybook = jest.requireActual('@storybook/angular');
 
   configure({ ...options, storybook });
 
   return {
     framework: 'angular' as const,
-    renderTree: require.requireActual('./renderTree').default,
+    renderTree: jest.requireActual('./renderTree').default,
     renderShallowTree: () => {
       throw new Error('Shallow renderer is not supported for angular');
     },