+++ app/react/src/client/index.ts
@@ -8,6 +8,7 @@ export {
   getStorybook,
   raw,
   forceReRender,
+  argsStory,
 } from './preview';
 
 if (module && module.hot && module.hot.decline) {