+++ docs/gatsby-config.js
@@ -27,6 +27,7 @@ module.exports = {
         '/guides/guide-riot/',
         '/guides/guide-svelte/',
         '/guides/guide-preact/',
+        '/guides/guide-web-components/',
       ],
       configurations: [
         '/configurations/overview/',