+++ examples/ember-cli/stories/addon-backgrounds.stories.js
@@ -4,10 +4,13 @@ export default {
   title: 'Addon/Backgrounds',
 
   parameters: {
-    backgrounds: [
-      { name: 'light', value: '#eeeeee' },
-      { name: 'dark', value: '#222222', default: true },
-    ],
+    backgrounds: {
+      default: 'dark',
+      values: [
+        { name: 'light', value: '#eeeeee' },
+        { name: 'dark', value: '#222222' },
+      ],
+    },
   },
 };